#29 Introduced the clipping.h file; moved all generic clipping code from graphics.
